wall_tracker_app/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
psycopg2-binary
pytest-django>=4.5,<5.0
pytest>=6.2,<7.0
gunicorn>=20.0,<21.0